requirements.txt
numpy
Pillow
scipy
matplotlib

